Adding item description and value to create an order

+++ PaymentSystem/Source/PaymentSystem.WebApi/Entities/Item.cs
@@ -5,12 +5,12 @@ namespace PaymentSystem.WebApi.Entities
     public class Item
     {
         public string title { get; set; }
-        public int unit_price { get; set; }
+        public double unit_price { get; set; }
         public int quantity { get; set; }
         public string unit_measure { get; set; }
-        public int total_amount { get; set; }
+        public double total_amount { get; set; }
 
-        public Item(string title, int unit_price, int quantity, string unit_measure, int total_amount)
+        public Item(string title, double unit_price, int quantity, string unit_measure, double total_amount)
         {
             this.title = title;
             this.unit_price = unit_price;
